Both branches independently hoisted MMR node-store/proof math into
@forestrie/merklelog; this merge converges them. receipt-verify now consumes
openMassifNodeStore / MassifNodeStore / inclusionProof / peakIndexForLeafProof /
massifIndexFromMMRIndex / peakMMRIndexes from @forestrie/merklelog (hoist
preserved) and uses the canonical deterministic CBOR codec
(encode/decodeCborDeterministic) at every call site.

cbor-x fully eradicated: zero code imports, no package.json/lockfile refs,
absent from node_modules. Conflicts resolved in build-receipt-offline.ts (+ its
tests), shared/encoding/src/index.ts, receipt-verify/src/index.ts, and
receipt-verify/package.json (0.5.0).

Verified: production packages typecheck clean; receipt-verify (24) + encoding
(103) tests pass incl. check:browser-safe. Pre-existing unrelated typecheck
fails in tests/canopy-api and libs/chain-rpc are untouched.

+++ packages/apps/canopy-api/src/scrapi/patch-grant-idtimestamp.ts
@@ -34,7 +34,7 @@ export async function patchGrantIdtimestamp(opts: {
 }): Promise<{ ok: true } | { ok: false; status: number; detail: string }> {
   const body = encodeCborDeterministic({
     idtimestamp: u64Be8(opts.idtimestamp),
-  }) as Uint8Array;
+  });
 
   const path = `/api/forest/${opts.rootLogId}/grants/${opts.subjectLogId}/idtimestamp`;
   let res: Response;
